+++ crates/e2e/sandbox/src/macros.rs
@@ -193,6 +193,8 @@ mod construct_runtime {
         type MaxCodeLen = ConstU32<{ 123 * 1024 }>;
         type MaxStorageKeyLen = ConstU32<128>;
         type UnsafeUnstableInterface = ConstBool<false>;
+        type UploadOrigin = $crate::frame_system::EnsureSigned<Self::AccountId>;
+        type InstantiateOrigin = $crate::frame_system::EnsureSigned<Self::AccountId>;
         type MaxDebugBufferLen = ConstU32<{ 2 * 1024 * 1024 }>;
         type Migrations = ();
         type DefaultDepositLimit = DefaultDepositLimit;